+++ src/route.rs
@@ -22,7 +22,7 @@ use crate::{
     assets::assets_handler,
     handler::{
         auth_handler::{login_page_handler, login_user_handler, logout_handler, session_check_handler},
-        bookmarks::{bookmark_content_handler, bookmark_create_handler, bookmark_new_handler, fetch_title_handler},
+        bookmarks::{bookmark_content_handler, bookmark_create_handler, bookmark_new_handler, scrape_site_handler},
         handle_404, home_handler,
         middlewares::auth_user_middleware,
         settings::{settings_handler, theme_toggle_handler, update_theme_handler},
@@ -90,7 +90,7 @@ fn create_router(app_state: Arc<AppState>) -> Router {
         .route("/api/bookmarks", get(bookmark_content_handler))
         .route("/api/tags", get(tag_list_handler))
         .route("/api/tags/autocomplete", get(tag_autocomplete_handler))
-        .route("/api/fetch-title", post(fetch_title_handler))
+        .route("/api/fetch-title", post(scrape_site_handler))
         .route("/api/settings/theme-toggle", get(theme_toggle_handler))
         .route("/api/settings/theme", post(update_theme_handler))
         .route("/api/session-check", get(session_check_handler))